Move cancel build to deployment endpoint
