+++ src/Appwrite/Platform/Modules/Sites/Http/Deployments/Status/Update.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Appwrite\Platform\Modules\Functions\Http\Deployments\Builds;
+namespace Appwrite\Platform\Modules\Sites\Http\Deployments\Status;
 
 use Appwrite\Event\Event;
 use Appwrite\Extend\Exception;
@@ -25,25 +25,24 @@ class Update extends Action
 
     public static function getName()
     {
-        return 'updateDeploymentBuild';
+        return 'updateDeploymentStatus';
     }
 
     public function __construct()
     {
         $this
             ->setHttpMethod(Action::HTTP_REQUEST_METHOD_PATCH)
-            ->setHttpPath('/v1/functions/:functionId/deployments/:deploymentId/build')
-            ->desc('Cancel deployment')
-            ->groups(['api', 'functions'])
-            ->label('scope', 'functions.write')
-            ->label('resourceType', RESOURCE_TYPE_FUNCTIONS)
+            ->setHttpPath('/v1/sites/:siteId/deployments/:deploymentId/status')
+            ->desc('Update deployment status')
+            ->groups(['api', 'sites'])
+            ->label('scope', 'sites.write')
             ->label('audits.event', 'deployment.update')
-            ->label('audits.resource', 'function/{request.functionId}')
+            ->label('audits.resource', 'site/{request.siteId}')
             ->label('sdk', new Method(
-                namespace: 'functions',
-                name: 'updateDeploymentBuild',
+                namespace: 'sites',
+                name: 'updateDeploymentStatus',
                 description: <<<EOT
-                Cancel an ongoing function deployment build. If the build is already in progress, it will be stopped and marked as canceled. If the build hasn't started yet, it will be marked as canceled without executing. You cannot cancel builds that have already completed (status 'ready') or failed. The response includes the final build status and details.
+                Cancel an ongoing site deployment build. If the build is already in progress, it will be stopped and marked as canceled. If the build hasn't started yet, it will be marked as canceled without executing. You cannot cancel builds that have already completed (status 'ready') or failed. The response includes the final build status and details.
                 EOT,
                 auth: [AuthType::KEY],
                 responses: [
@@ -53,7 +52,7 @@ class Update extends Action
                     )
                 ]
             ))
-            ->param('functionId', '', new UID(), 'Function ID.')
+            ->param('siteId', '', new UID(), 'Site ID.')
             ->param('deploymentId', '', new UID(), 'Deployment ID.')
             ->inject('response')
             ->inject('dbForProject')
@@ -62,12 +61,12 @@ class Update extends Action
             ->callback([$this, 'action']);
     }
 
-    public function action(string $functionId, string $deploymentId, Response $response, Database $dbForProject, Document $project, Event $queueForEvents)
+    public function action(string $siteId, string $deploymentId, Response $response, Database $dbForProject, Document $project, Event $queueForEvents)
     {
-        $function = $dbForProject->getDocument('functions', $functionId);
+        $site = $dbForProject->getDocument('sites', $siteId);
 
-        if ($function->isEmpty()) {
-            throw new Exception(Exception::FUNCTION_NOT_FOUND);
+        if ($site->isEmpty()) {
+            throw new Exception(Exception::SITE_NOT_FOUND);
         }
 
         $deployment = $dbForProject->getDocument('deployments', $deploymentId);
@@ -88,7 +87,7 @@ class Update extends Action
                 'deploymentId' => $deployment->getId(),
                 'status' => 'canceled',
                 'path' => '',
-                'runtime' => $function->getAttribute('runtime'),
+                'runtime' => $site->getAttribute('framework'),
                 'source' => $deployment->getAttribute('path', ''),
                 'sourceType' => '',
                 'logs' => '',
@@ -128,7 +127,7 @@ class Update extends Action
         }
 
         $queueForEvents
-            ->setParam('functionId', $function->getId())
+            ->setParam('siteId', $site->getId())
             ->setParam('deploymentId', $deployment->getId());
 
         $response->dynamic($build, Response::MODEL_BUILD);
